refactor: 티켓 예매 ViewBinder 제거
